Add last update code and skip already up-to-date extensions

+++ bin/update.rb
@@ -10,7 +10,9 @@ if !File.directory?(mediawiki_git_repo)
 end
 
 push_command = 'git subtree push --prefix=extensions/%s mediawiki-extensions %s'
+commit_command = 'git log -n1 --pretty=format:"%an::%ai::%s"'
 FileUtils.cd(mediawiki_git_repo) do
+  system('git fetch mediawiki-extensions')
   system('git svn rebase')
 
   entries = Dir.entries('extensions/')
@@ -21,6 +23,10 @@ FileUtils.cd(mediawiki_git_repo) do
       next
     end
 
+    current_commit = `#{commit_command} mediawiki-extensions/#{filename}`
+    latest_commit = `#{commit_command} -- extensions/#{filename}`
+    next if current_commit == latest_commit
+
     system(push_command % [filename, filename])
     exit unless $?.success?
     system("rm -rf .git/subtree-cache/#{$?.pid}")
@@ -30,6 +36,12 @@ FileUtils.cd(mediawiki_git_repo) do
 end
 
 FileUtils.cd("#{File.dirname(__FILE__)}/..") do
+  File.open('LAST_UPDATED_AT', 'w') do |file|
+    file.puts Time.now.inspect
+  end
+
+  system('git commit -m "Extensions updated" LAST_UPDATED_AT')
+  system('git push origin master')
   system('git fetch')
   system('git gc --aggressive')
 end
